+++ apps/dialtone-documentation/docs/.vuepress/theme/client.js
@@ -9,7 +9,7 @@ import { flushPromises } from '@workspaceRoot/common/utils';
 
 // CSS
 import '@dialpad/dialtone-css/lib/dist/dialtone.css';
-import '@dialpad/dialtone/vue3/css';
+import '@dialpad/dialtone-vue/css';
 import './assets/less/dialtone-docs.less';
 import './assets/less/dialtone-syntax.less';