+++ packages/dialtone-vue2/components/emoji_picker/modules/emoji_selector.vue
@@ -148,6 +148,7 @@ export default {
 
   data () {
     return {
+      tabLabelsRefs: [],
       emojiRefs: [],
       emojiFilteredRefs: [],
       isFiltering: false,
@@ -184,8 +185,8 @@ export default {
 
     tabLabels () {
       return this.recentlyUsedEmojis.length
-        ? this.tabSetLabels.map((label, index) => ({ label, ref: this.$refs[`tabLabelRef-${index}`] }))
-        : this.tabSetLabels.slice(1).map((label, index) => ({ label, ref: this.$refs[`tabLabelRef-${index}`] }));
+        ? this.tabSetLabels.map((label) => ({ label }))
+        : this.tabSetLabels.slice(1).map((label) => ({ label }));
     },
 
     tabs () {
@@ -254,10 +255,10 @@ export default {
 
   methods: {
     setupTabLabelRefs () {
-      this.tabSetLabels?.forEach((label, index) => {
+      this.tabSetLabels?.forEach((_, index) => {
         const refKey = `tabLabelRef-${index}`;
         if (this.$refs[refKey]) {
-          this.$set(this.tabLabels, index, { label, ref: this.$refs[refKey] });
+          this.$set(this.tabLabelsRefs, index, { ref: this.$refs[refKey] });
         }
       });
     },
@@ -329,31 +330,11 @@ export default {
     scrollToTab: function (tabIndex, focusFirstEmoji) {
       const vm = this;
       if (focusFirstEmoji === undefined) { focusFirstEmoji = true; }
-      const tabLabel = vm.tabLabels[tabIndex - 1];
-      const tabElement = tabLabel.ref[0];
+      const tabElement = vm.tabLabelsRefs[tabIndex - 1].ref[0];
 
       vm.$nextTick(function () {
         const container = vm.$refs.listRef;
-        const offsetTop = tabIndex === '1' ? 0 : tabElement.offsetTop - 20;
-
-        let isScrolling = true;
-        let prevScrollTop = container.scrollTop;
-        vm.$emit('is-scrolling', true);
-
-        /* eslint-disable-next-line complexity */
-        container.addEventListener('scroll', function () {
-          if (isScrolling) {
-            const scrollTop = container.scrollTop;
-            if (
-              (prevScrollTop < scrollTop && scrollTop >= offsetTop) ||
-            (prevScrollTop > scrollTop && scrollTop <= offsetTop)
-            ) {
-              isScrolling = false;
-              vm.$emit('is-scrolling', false);
-            }
-            prevScrollTop = scrollTop;
-          }
-        });
+        const offsetTop = tabIndex === 1 ? 0 : tabElement.offsetTop - 15;
 
         container.scrollTop = offsetTop;
 
@@ -471,7 +452,7 @@ export default {
         this.handleHorizontalNavigation('right', indexTab, indexEmoji);
       }
 
-      if (event.key === 'Tab') {
+      if (event.key === 'Tab' && !event.shiftKey) {
         if (this.focusEmoji(indexTab + 1, 0)) {
           this.scrollToTab((indexTab + 1) + 1, false);
         } else {
@@ -591,7 +572,6 @@ export default {
 
     setTabLabelObserver () {
       this.tabLabelObserver = new IntersectionObserver(entries => {
-        this.$emit('is-scrolling', false);
         /* eslint-disable-next-line complexity */
         entries.forEach(entry => {
           const { target } = entry;
@@ -619,7 +599,7 @@ export default {
     },
 
     focusLastEmoji () {
-      this.focusEmoji(this.tabs.length - 1, 0);
+      this.scrollToTab(this.tabs.length, true);
     },
 
   },