+++ packages/dialtone-vue2/components/rich_text_editor/rich_text_editor.test.js
@@ -52,6 +52,7 @@ describe('DtRichTextEditor tests', () => {
   afterEach(() => {
     mockProps = {};
     mockListeners = {};
+    wrapper.destroy();
   });
 
   describe('Presentation Tests', () => {