+++ packages/dialtone-vue2/recipes/buttons/callbar_button/callbar_button.test.js
@@ -37,7 +37,6 @@ describe('DtRecipeCallbarButton Tests', () => {
       slots,
       provide,
       listeners,
-      attachTo: document.body,
       localVue: testContext.localVue,
     });
     _setChildWrappers();