+++ packages/dialtone-vue2/recipes/conversation_view/message_input/message_input.test.js
@@ -54,7 +54,6 @@ const _mountWrapper = () => {
     listeners,
     slots,
     localVue: testContext.localVue,
-    attachTo: document.body,
   });
 };